fix: Relocation, SaveWeight API 파일 분리
