+++ relocation-service/src/main/kotlin/com/baro/relocation/controller/RelocationController.kt
@@ -2,8 +2,6 @@ package com.baro.relocation.controller
 
 import com.baro.relocation.dto.RelocationAssignRequest
 import com.baro.relocation.dto.RelocationAssignResponse
-import com.baro.relocation.dto.StandWeightRequest
-import com.baro.relocation.dto.StandWeightResponse
 import com.baro.relocation.service.RelocationService
 import io.swagger.v3.oas.annotations.Operation
 import org.springframework.http.ResponseEntity
@@ -13,26 +11,11 @@ import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 
 @RestController
+@RequestMapping("/relocation")
 class RelocationController (
     private val relocationService: RelocationService
 ) {
-    @PostMapping("/internal/relocation/standWeights")
-    @Operation(summary = "가중치 데이터 수신", description = "private에서 산출한 가중치 수신 / 저장")
-    fun receiveWeights(
-        @RequestBody request: StandWeightRequest
-    ): ResponseEntity<StandWeightResponse> {
-
-        relocationService.saveWeights(request)
-
-        return ResponseEntity.ok(
-            StandWeightResponse(
-                message = "가중치 저장 완료",
-                count = request.weights.size
-            )
-        )
-    }
-
-    @PostMapping("/relocation/assign")
+    @PostMapping("/assign")
     @Operation(summary = "차량 재배치 택시 정거장 할당", description = "운행이 끝난 차량들 재배치할 택시 정거장 위치 반환")
     fun assignRelocation(
         @RequestBody request: RelocationAssignRequest
