+++ relocation-service/src/test/kotlin/com/baro/relocation/SaveWeightServiceApplicationTests.kt
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
 
 @SpringBootTest
-class RelocationServiceApplicationTests {
+class SaveWeightServiceApplicationTests {
     @Test
     fun `스프링 컨텍스트가 정상적으로 로드된다`() {
     }
